requirements.txt
groq
fastapi
uvicorn
python-dotenv
requests
